message_bus_lib/requirements.txt
# Azure messaging and identity
azure-servicebus==7.14.2
azure-identity==1.21.0

setuptools==80.0.0
